Adding unit test for FireCloudClient#api_status
New unit test to check that the service endpoints that SCP relies on in FireCloud (Rawls, Agora, Sam, and Thurloe) are present in api_status responses.  If any of these services are renamed/removed, then we will know we have a regression that will have major functionality implications moving forward as we rely on these services by name.