refactoring can_download?, fixing bug with read-only access shares
study.can_download?(user) now calls study.has_bucket_access?(user) which is a helper that checks for explicit permissions in FireCloud (specifically, READER or WRITER).  This makes can_download? cleaner as it can check if the user is signed in and the study is public to return quickly, or drill deeper to examine permissions.

Also fixing a bug that causes study.set_read_only_access to not remove StudyShare objects when removing read-only access to studies (for public data).  Now creating/destroying shares directly and letting validation callbacks handle setting the permissions in FireCloud, which is what should have been done in the first place.