auth refactoring for FireCloudClient and Api/V1/SiteController
Resolved lingering issue of setting the access_token for a user-scoped instance of FireCloudClient when calling via REST API.  Resolution is to apply the incoming API access token into the client and set an expiration date of 5 minutes, as each client is instantiated for one request only, so renewing the token is irrelevant - the user must supply a valid OAuth2 token.
Also ensured that api_access_tokens are checked for expiration on all requests, and responding correctly with 401s where appropriate.