Adding permission description for shares, other fixes
Adding the permission description to the list of study shares on the 'Show Details' study page.
Fixing a regression with the download refactoring and generating api_urls.
Adding method snapshot versions to the list of available workflows in Analysis tab.
Updating text on the 'Update' button on AnalysisConfiguration configure page.