adding csrf protection to omniauth/google_oauth2 sign in routes/links
As per CVE-2015-9284, Omniauth-generated sign in routes are subject to CSRF forgery and can allow secondary accounts to be linked to primary accounts without user interaction.  While the impact of this is very low for SCP (we only use Google OAuth2, and do not have other accounts), the fix removes the vulnerability altogether.  SCP will respond with 404 on GETs to the sign_in path, and 422 to any POSTs that do not have a valid CSRF token.