refactoring for readability, fixing search submission, caching genes
Refactored methods for readability re: PR comments.  Also, adding close() event to autocomplete to enable submitting search terms with ENTER key even if a user did not select an element from the list as select() would not have fired.  Unique genes for a study are cached in the window to reduce the number of queries to Firestore that need to be run in order to drive autocomplete.