adding boot_disk_size_gb to PapiClient#create_virtual_machine_object
Adding parameter to adjust the size of the boot disk attached to all ingest pipeline runs.  For now, this is hard-coded to 100GB, but in the future this could be abstracted out to a configuration setting (along with machine_type) to allow for dynamic control of machines/disks in ingest pipeline runs.