Addressing issue with blank/non-file analysis_parameters
Fixes an issue with output parameters who's data type includes a '?' which causes nil values to be passed if the parameter is not configured by an admin.  Now ignoring non-file outputs when syncing if no file can be resolved using the value of the output.  Also, small UI updates to make pipeline names include namespace for disambiguation.